requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
opencv-python
numpy
matplotlib
imutils
deepface
pathlib
scikit-learn
supabase
sendgrid
twilio
requests
maileroo
psycopg2-binary
flask-migrate
python-dotenv
python-decouple